Create RegisterPage in authSlice branch, rename RegisterForm to AuthForm and move to separate file, log token to console
