+++ client/components/RegisterPage.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import AuthForm from "./AuthForm";
+
+const RegisterPage = () => {
+    return(
+        <>
+            <div>I am the Register Page</div>
+            <AuthForm />
+        </>
+    )
+}
+
+export default RegisterPage;
