+++ package.json
@@ -25,7 +25,8 @@
     "prisma": "^5.4.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-redux": "^8.1.3"
+    "react-redux": "^8.1.3",
+    "react-router-dom": "^6.18.0"
   },
   "devDependencies": {
     "@types/express": "^4.17.19",
